Move CustomerInvoice and Invoice classes outside Import namespace
